Centralize kotlin compiler configuration for gradle/gradle build
