+++ build-logic/basics/src/main/kotlin/gradlebuild/basics/util/KotlinSourceParser.kt
@@ -16,8 +16,9 @@
 
 package gradlebuild.basics.util
 
-import org.gradle.internal.jvm.Jvm
 
+import gradlebuild.basics.kotlindsl.configureKotlinCompilerForGradleBuild
+import org.gradle.internal.jvm.Jvm
 import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
 import org.jetbrains.kotlin.cli.common.config.addKotlinSourceRoots
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
@@ -31,19 +32,8 @@ import org.jetbrains.kotlin.com.intellij.openapi.util.Disposer
 import org.jetbrains.kotlin.config.CommonConfigurationKeys
 import org.jetbrains.kotlin.config.CompilerConfiguration
 import org.jetbrains.kotlin.config.JVMConfigurationKeys
-import org.jetbrains.kotlin.config.ApiVersion
-import org.jetbrains.kotlin.config.JvmAnalysisFlags
-import org.jetbrains.kotlin.config.JvmDefaultMode
-import org.jetbrains.kotlin.config.JvmTarget
-import org.jetbrains.kotlin.config.LanguageVersion
-import org.jetbrains.kotlin.config.LanguageVersionSettingsImpl
 import org.jetbrains.kotlin.psi.KtFile
-import org.jetbrains.kotlin.load.java.JavaTypeEnhancementState
-import org.jetbrains.kotlin.load.java.Jsr305Settings
-import org.jetbrains.kotlin.load.java.ReportLevel
 import org.jetbrains.kotlin.utils.PathUtil
-
-
 import java.io.File
 
 
@@ -106,27 +96,6 @@ class KotlinSourceParser {
     private
     fun configureKotlinCompilerIoForWindowsSupport() =
         org.jetbrains.kotlin.cli.common.environment.setIdeaIoUseFallback()
-
-    private
-    fun CompilerConfiguration.configureKotlinCompilerForGradleBuild() {
-
-        put(
-            CommonConfigurationKeys.LANGUAGE_VERSION_SETTINGS,
-            LanguageVersionSettingsImpl(
-                languageVersion = LanguageVersion.KOTLIN_1_4,
-                apiVersion = ApiVersion.KOTLIN_1_4,
-                analysisFlags = mapOf(
-                    JvmAnalysisFlags.javaTypeEnhancementState to JavaTypeEnhancementState(
-                        Jsr305Settings(ReportLevel.STRICT, ReportLevel.STRICT)
-                    ) { ReportLevel.STRICT },
-                    JvmAnalysisFlags.jvmDefaultMode to JvmDefaultMode.ENABLE
-                )
-            )
-        )
-
-        put(JVMConfigurationKeys.PARAMETERS_METADATA, true)
-        put(JVMConfigurationKeys.JVM_TARGET, JvmTarget.JVM_1_8)
-    }
 }
