+++ build-logic/uber-plugins/src/main/kotlin/gradlebuild.kotlin-library.gradle.kts
@@ -15,9 +15,8 @@
  */
 
 import gradlebuild.basics.accessors.kotlin
+import gradlebuild.basics.kotlindsl.configureKotlinCompilerForGradleBuild
 import org.gradle.api.internal.initialization.DefaultClassLoaderScope
-import org.jetbrains.kotlin.gradle.dsl.JvmTarget
-import org.jetbrains.kotlin.gradle.dsl.KotlinVersion
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 
@@ -70,21 +69,6 @@ tasks {
         systemProperty(
             DefaultClassLoaderScope.STRICT_MODE_PROPERTY,
             true
-        )
-    }
-}
-
-fun KotlinCompile.configureKotlinCompilerForGradleBuild() {
-    compilerOptions {
-        allWarningsAsErrors = true
-        apiVersion = KotlinVersion.KOTLIN_1_8
-        languageVersion = KotlinVersion.KOTLIN_1_8
-        jvmTarget = JvmTarget.JVM_1_8
-        freeCompilerArgs.addAll(
-            "-Xjsr305=strict",
-            "-java-parameters",
-            "-Xsam-conversions=class",
-            "-Xskip-metadata-version-check",
         )
     }
 }
